// src/enums/CustomResponseType.ts
export enum CustomResponseType {
  SUCCESS = '成功',
  SYSTEM_ERROR = '系統錯誤',
  VALIDATION_ERROR = '驗證錯誤',
  NOT_FOUND = '找不到',
  UNAUTHORIZED = '未授權',
  FORBIDDEN = '禁止',
  BAD_REQUEST = '錯誤請求',
  CONFLICT = '衝突',
  CREATED = '已創建',
  UPDATED = '已更新',
  DELETED = '已刪除',
  DATABASE_OPERATION_FAILED = '資料庫操作失敗',
}